requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.1
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
branca==0.8.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
folium==0.19.5
fonttools==4.57.0
frozenlist==1.6.0
gast==0.6.0
geopandas==1.0.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
graphviz==0.21
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaleido==0.2.1
keras==3.9.1
kiwisolver==1.4.8
libclang==18.1.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
multidict==6.4.3
namex==0.0.8
narwhals==1.34.1
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.1.3
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
packaging==24.2
Panda3D==1.10.15
pandas==2.3.0
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.0
pydantic_core==2.33.0
pydeck==0.9.1
pydyf==0.11.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
PyMsgBox==1.0.9
pyodbc==5.2.0
pyogrio==0.11.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.9.0
pyphen==0.17.2
pypiwin32==223
pyproj==3.7.1
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PySimpleGUI==5.0.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
setuptools==78.1.0
shapely==2.1.0
shiboken6==6.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
streamlit==1.44.1
streamlit-drawable-canvas==0.9.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
ttkbootstrap==1.13.8
twilio==9.6.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.35.0
virtualenv==20.30.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasyprint==65.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xlrd==2.0.1
xyzservices==2025.4.0
yarl==1.20.0
zopfli==0.2.3.post1